feat: implement built-in AI signal handling and streaming functionality
- Add signal handling utilities for aborting operations with `abortError`, `mergeSignals`, and `raceAbort`.
- Create `streamChunks` function to yield chunks from a ReadableStream with abort support.
- Introduce tests for streaming functionality to ensure correct chunk yielding and abort behavior.
- Implement lifecycle management for built-in AI namespaces with `useLifecycle`, handling state transitions and user activation.
- Add support for shared context management in local storage with `useSharedContext`.
- Define built-in AI namespaces and a utility function `isSupported` to check for global namespace availability.
- Create mock instances for testing various AI functionalities, including translators, summarizers, and proofreaders.
