+++ src/shared/built-in-ai/errors.test.ts
