+++ src/shared/built-in-ai/errors.ts
