+++ src/shared/built-in-ai/hooks/useLanguageDetector.ts
