+++ src/shared/built-in-ai/hooks/useRewriter.ts
