+++ src/shared/built-in-ai/hooks/useSummarizer.ts
