+++ src/shared/built-in-ai/hooks/useTranslator.ts
