+++ src/shared/built-in-ai/hooks/useWriter.ts
