+++ src/shared/built-in-ai/internal/activation.ts
