+++ src/shared/built-in-ai/internal/options-equality.ts
