+++ src/shared/built-in-ai/internal/test-fakes/ai-namespace-fake.ts
