+++ src/shared/built-in-ai/internal/test-fakes/instance-fakes.ts
