+++ src/shared/built-in-ai/internal/types.ts
