+++ src/shared/built-in-ai/namespaces.test.ts
