+++ src/shared/built-in-ai/namespaces.ts
