+++ src/shared/built-in-ai/sharedContext.ts
